[piechart] - add FilterArticleDto - calculate aggregation with jpa - pagination fixed

+++ src/main/java/mk/ukim/finki/web/financeproject/model/piechart/Dataset.java
@@ -8,7 +8,7 @@ import java.util.List;
 @Getter
 @Setter
 public class Dataset {
-    private List<Integer> data;
+    private List<Long> data;
     private List<String> backgroundColor;
     private List<String> hoverBackgroundColor;
 }
